requirements.txt
# Core ML/NLP Libraries
transformers>=4.35.0
datasets>=2.14.0
torch>=2.0.0
sentencepiece>=0.1.99
accelerate>=0.24.0

# Evaluation Metrics
rouge-score>=0.1.2
nltk>=3.8.1
evaluate>=0.4.0

# Configuration Management
pyyaml>=6.0
omegaconf>=2.3.0

# Logging and Tracking
tensorboard>=2.14.0
wandb>=0.15.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0

# Utilities
python-dotenv>=1.0.0
rich>=13.5.0
